# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking API",
        "description": "Documentation de l'API de gestion des parkings",
        "version": "1.0"
    },
    "paths": {
        "/api/parkings": {
            "get": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Lister tous les parkings",
                "operationId": "906b203d7aa2fe58b5c3c0a767f95b8d",
                "responses": {
                    "200": {
                        "description": "Liste des parkings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Parking"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Ajouter un parking",
                "operationId": "e695faaf3f2d6470ac8c87f62b5221c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "total_spaces"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parking Central"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris, France"
                                    },
                                    "total_spaces": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking ajouté avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/parkings/{id}": {
            "get": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Obtenir les détails d'un parking",
                "operationId": "e52a76d93475b17397e56b9c3a9a1507",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du parking"
                    },
                    "404": {
                        "description": "Parking non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Mettre à jour un parking",
                "operationId": "1d51b081ba742017d7d1401ef3984091",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "total_spaces"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Parking Central"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Paris, France"
                                    },
                                    "total_spaces": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parking mis à jour"
                    },
                    "404": {
                        "description": "Parking non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Supprimer un parking",
                "operationId": "c9a41b7816141be701626d0e88a0d614",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du parking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parking supprimé"
                    },
                    "404": {
                        "description": "Parking non trouvé"
                    }
                }
            }
        },
        "/api/parkings/search": {
            "get": {
                "tags": [
                    "Parkings"
                ],
                "summary": "Rechercher un parking par emplacement",
                "operationId": "65a84f139d47b1e9394733f4ea6d8498",
                "parameters": [
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Emplacement recherché",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des parkings trouvés"
                    },
                    "404": {
                        "description": "Aucun parking trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Parking": {
                "title": "Parking",
                "description": "Modèle d'un parking",
                "required": [
                    "name",
                    "location",
                    "total_spaces"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Parking Central"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris, France"
                    },
                    "total_spaces": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-13T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-13T12:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}